refactor test into unit and e2e.
