+++ jest.config.e2e.js
@@ -1,0 +1,7 @@
+// eslint-disable-next-line
+import commonConfig from './jest.config.js';
+
+export default {
+  ...commonConfig,
+  testMatch: ['**/*.e2e.test.{js,ts}']
+}
